Remove overlay in favor of background color + scrollbar dim
