+++ source/cli/cli.tsx
@@ -33,7 +33,6 @@ import type { LoadedSession, Session } from "../session-history/index.ts";
 import { useAppStore } from "../state.ts";
 import { FOREGROUND_COLOR, THEME_COLOR } from "../theme.ts";
 import { KeyboardProvider } from "../hooks/use-keyboard.ts";
-import { AppFocusOverlay } from "../components/app-focus-overlay.tsx";
 import { render, type CreateRootOptions } from "paintcannon-react";
 const __dirname = import.meta.dirname;
 const INTERACTIVE_RENDER_OPTIONS = {
@@ -42,12 +41,7 @@ const INTERACTIVE_RENDER_OPTIONS = {
   captureCtrlC: true,
 } satisfies CreateRootOptions;
 function renderInteractive(element: React.ReactNode) {
-  const root = render(
-    <KeyboardProvider>
-      <AppFocusOverlay>{element}</AppFocusOverlay>
-    </KeyboardProvider>,
-    INTERACTIVE_RENDER_OPTIONS,
-  );
+  const root = render(<KeyboardProvider>{element}</KeyboardProvider>, INTERACTIVE_RENDER_OPTIONS);
   root.container.style.position = "relative";
   root.container.style.color = FOREGROUND_COLOR;
   return root;
