+++ source/theme.ts
@@ -8,10 +8,11 @@ export const DIFF_HEADER_COLOR = "gray";
 export const DIFF_MARKER_COLOR = "black";
 export const CODE_GUTTER_COLOR = "gray";
 export const BACKGROUND_COLOR = "#0f172a";
+export const DIMMED_BACKGROUND_COLOR = "#020617";
 export const FOREGROUND_COLOR = "#a3a3a3";
 export const SCROLLBAR_COLOR = "#64748b #1e293b";
+export const DIMMED_SCROLLBAR_COLOR = "#475569 #0f172a";
 export const SUBTLE_SCROLLBAR_COLOR = "#475569 #1e293b";
-export const APP_OVERLAY_Z_INDEX = 2147483647;
 export const MARKDOWN_BLOCKQUOTE_BORDER_COLOR = "gray";
 export const MARKDOWN_CODE_BLOCK_BORDER_COLOR = "gray";
 export const MARKDOWN_HEADING_COLORS = [
